Fix incorrect fan percentage
Changed line 69 to show static speed of about 75% instead of 30%.

I was perusing through your other project for idrac reverse engineering and ran across a section about the same feature here: https://github.com/Fohdeesha/idrac-7-8-reverse-engineering/tree/master/squashfs-ipmi#thermal--fans .  You had found the fan speed was actually 75% with an unknown PCI-e card.  I believe 75% is likely the more accurate reading.  The 30% I was reading from was in the sensors menu and that might be what it would set it to if the fan speeds weren't static.