change reset to boot_primary
I have seen on a few switches (4 so far) that they come with the secondary flash as default boot. Even after doing factory set-default it will still boot the secondary flash after updating from the bootloader. The switch models I've experienced this on were the ICX7150 and ICX7250. Since the guide flashes the primary we can just guide people to boot the primary so there is no confusion in case it boots from secondary.